+++ djangocms_url_manager/cms_config.py
@@ -9,20 +9,51 @@ from djangocms_url_manager.utils import get_page_search_results, parse_settings
 from .models import Url
 
 
+try:
+    from djangocms_versioning.constants import DRAFT  # NOQA
+
+    djangocms_versioning_installed = True
+except ImportError:
+    djangocms_versioning_installed = False
+
+try:
+    from djangocms_moderation import __version__  # NOQA
+
+    djangocms_moderation_installed = True
+except ImportError:
+    djangocms_moderation_installed = False
+
+
 class UrlCMSAppConfig(CMSAppConfig):
     djangocms_url_manager_enabled = True
-    djangocms_versioning_enabled = True
     url_manager_supported_models = [Page]
+    url_manager_supported_models_search_helpers = {
+        Page: get_page_search_results,
+    }
+    # djangocms-navigation settings
     djangocms_navigation_enabled = getattr(
         settings, "DJANGOCMS_NAVIGATION_CMS_MODELS_ENABLED", False
     )
     navigation_models = {Url: ["internal_name"]}
+    # djangocms-moderation settings
+    djangocms_moderation_enabled = getattr(
+        settings, 'MODERATING_URL_MANAGER_MODELS_ENABLED', True
+    )
+    if djangocms_moderation_enabled and djangocms_moderation_installed:
+        moderated_models = [Url]
+    # djangocms-versioning settings
+    djangocms_versioning_enabled = getattr(
+        settings, 'VERSIONING_URL_MANAGER_MODELS_ENABLED', True)
 
-    url_manager_supported_models_search_helpers = {
-        Page: get_page_search_results,
-    }
-
-    versioning = []
+    if djangocms_versioning_enabled and djangocms_versioning_installed:
+        from djangocms_versioning.datastructures import VersionableItem, default_copy
+        versioning = [
+            VersionableItem(
+                content_model=Url,
+                grouper_field_name='url_grouper',
+                copy_function=default_copy,
+            ),
+        ]
 
 
 class UrlManagerCMSExtension(CMSAppExtension):